add configs with their models (#421)
* add configs with their models

* fix tests

* doc update

* doc update

* fix path
