--- docs/source/package_reference/model_config.mdx
@@ -1,10 +1,0 @@
-# ModelConfig
-
-[[autodoc]] models.model_config.BaseModelConfig
-
-[[autodoc]] models.model_config.AdapterModelConfig
-[[autodoc]] models.model_config.DeltaModelConfig
-[[autodoc]] models.model_config.InferenceEndpointModelConfig
-[[autodoc]] models.model_config.InferenceModelConfig
-[[autodoc]] models.model_config.TGIModelConfig
-[[autodoc]] models.model_config.VLLMModelConfig
